Refactor away classic menu creation code
